# Benchmark default version profiling