--- a/lint/schema.go
+++ b/lint/schema.go
@@ -98,6 +98,9 @@ const confSchema = `{
 			"pool_size": {
 				"type": "integer"
 			},
+			"records_buffer_size": {
+				"type": "integer"
+			},
 			"storage_expiration_time": {
 				"type": "integer"
 			},