--- a/main.go
+++ b/main.go
@@ -166,7 +166,7 @@ func setupGlobals() {
 
 		analyticsStore := storage.RedisCluster{KeyPrefix: "analytics-"}
 		analytics.Store = &analyticsStore
-		analytics.Init()
+		analytics.Init(globalConf)
 
 		redisPurgeOnce.Do(func() {
 			store := storage.RedisCluster{KeyPrefix: "analytics-"}
@@ -1050,6 +1050,12 @@ func main() {
 
 	mainLog.Info("Stop signal received.")
 
+	// stop analytics workers
+	if config.Global().EnableAnalytics && analytics.Store == nil {
+		analytics.Stop()
+	}
+
+	// write pprof profiles
 	writeProfiles()
 
 	if config.Global().UseDBAppConfigs {