--- a/storage/redis_cluster.go
+++ b/storage/redis_cluster.go
@@ -558,13 +558,38 @@ func (r RedisCluster) GetAndDeleteSet(keyName string) []interface{} {
 }
 
 func (r RedisCluster) AppendToSet(keyName, value string) {
-	log.Debug("Pushing to raw key list: ", keyName)
-	log.Debug("Appending to fixed key list: ", r.fixKey(keyName))
+	fixedKey := r.fixKey(keyName)
+	log.WithField("keyName", keyName).Debug("Pushing to raw key list")
+	log.WithField("fixedKey", fixedKey).Debug("Appending to fixed key list")
 	r.ensureConnection()
-	_, err := r.singleton().Do("RPUSH", r.fixKey(keyName), value)
+	if _, err := r.singleton().Do("RPUSH", fixedKey, value); err != nil {
+		log.WithError(err).Error("Error trying to append to set keys")
+	}
+}
 
-	if err != nil {
-		log.Error("Error trying to delete keys: ", err)
+func (r RedisCluster) AppendToSetPipelined(key string, values []string) {
+	if len(values) == 0 {
+		return
+	}
+
+	fixedKey := r.fixKey(key)
+
+	// prepare pipeline data
+	pipeLine := make([]rediscluster.ClusterTransaction, len(values))
+	for index, val := range values {
+		pipeLine[index] = rediscluster.ClusterTransaction{
+			Cmd: "RPUSH",
+			Args: []interface{}{
+				fixedKey,
+				val,
+			},
+		}
+	}
+
+	// send pipelined command to Redis
+	r.ensureConnection()
+	if _, err := r.singleton().DoPipeline(pipeLine); err != nil {
+		log.WithError(err).Error("Error trying to append to set keys")
 	}
 }