@buger 2018-06-02
What if we have some moderate load, like 5 RPS, which does not hit timeout, appends to the buffer, and waits till it will fill. How big analytics delay can be in the worst situation?