@matiasinsaurralde 2018-06-04
Looks good and agree with Leon's comments, will rebase #1690 with this after the merge to see if more improvements show up